Use credentials in Dataset. Move debug_print method to utils
